Conflicts:
	starcluster/cluster.py

+++ starcluster/cluster.py
@@ -5,10 +5,11 @@ import string
 import pprint
 import warnings
 
+import iptools
+
 from starcluster import utils
 from starcluster import static
 from starcluster import spinner
-from starcluster import iptools
 from starcluster import sshutils
 from starcluster import managers
 from starcluster import userdata
@@ -285,18 +286,22 @@ class ClusterManager(managers.Manager):
             print 'Uptime: %s' % uptime
             print 'Zone: %s' % getattr(n, 'placement', 'N/A')
             print 'Keypair: %s' % getattr(n, 'key_name', 'N/A')
-            ebs_nodes = [n for n in nodes if n.attached_vols]
-            if ebs_nodes:
+            ebs_vols = []
+            for node in nodes:
+                devices = node.attached_vols
+                if not devices:
+                    continue
+                node_id = node.alias or node.id
+                for dev in devices:
+                    d = devices.get(dev)
+                    vol_id = d.volume_id
+                    status = d.status
+                    ebs_vols.append((vol_id, node_id, dev, status))
+            if ebs_vols:
                 print 'EBS volumes:'
-                for node in ebs_nodes:
-                    devices = node.attached_vols
-                    node_id = node.alias or node.id
-                    for dev in devices:
-                        d = devices.get(dev)
-                        vol_id = d.volume_id
-                        status = d.status
-                        print('    %s on %s:%s (status: %s)' %
-                              (vol_id, node_id, dev, status))
+                for vid, nid, dev, status in ebs_vols:
+                    print('    %s on %s:%s (status: %s)' %
+                          (vid, nid, dev, status))
             else:
                 print 'EBS volumes: N/A'
             spot_reqs = cl.spot_requests
@@ -679,9 +684,13 @@ class Cluster(object):
         # TODO:  problem here is that it interrogates all nodes in the
         # TODO:  security group, not just the starcluster nodes that
         # TODO:  we're interested in
-        filters = {'group-name': self._security_group,
-                   'instance-state-name': states,
+        filters = {'instance-state-name': states,
                    'tag:starcluster': 'starcluster',}
+        cluster_group = self.cluster_group
+        if cluster_group.vpc_id:
+            filters['instance.group-name'] = cluster_group.name
+        else:
+            filters['group-name'] = cluster_group.name
         nodes = self.ec2.get_all_instances(filters=filters)
         # remove any cached nodes not in the current node list from EC2
         current_ids = [n.id for n in nodes]
@@ -1322,6 +1331,7 @@ class Cluster(object):
         """
         Attach each volume to the master node
         """
+        wait_for_volumes = []
         for vol in self.volumes:
             volume = self.volumes.get(vol)
             device = volume.get('device')
@@ -1339,6 +1349,8 @@ class Cluster(object):
                      (vol.id, device))
             resp = vol.attach(self.master_node.id, device)
             log.debug("resp = %s" % resp)
+            wait_for_volumes.append(vol)
+        for vol in wait_for_volumes:
             self.ec2.wait_for_volume(vol, state='attached')
 
     def detach_volumes(self):
@@ -1421,17 +1433,19 @@ class Cluster(object):
             if spot.state not in ['cancelled', 'closed']:
                 log.info("Canceling spot instance request: %s" % spot.id)
                 spot.cancel()
-        sg = self.ec2.get_group_or_none(self._security_group)
-        pg = self.ec2.get_placement_group_or_none(self._security_group)
         s = self.get_spinner("Waiting for cluster to terminate...")
         try:
             while not self.is_cluster_terminated():
                 time.sleep(5)
         finally:
             s.stop()
-        if pg:
-            log.info("Removing %s placement group" % pg.name)
-            pg.delete()
+        region = self.ec2.region.name
+        if region in static.CLUSTER_REGIONS:
+            pg = self.ec2.get_placement_group_or_none(self._security_group)
+            if pg:
+                log.info("Removing %s placement group" % pg.name)
+                pg.delete()
+        sg = self.ec2.get_group_or_none(self._security_group)
         if sg:
             log.info("Removing %s security group" % sg.name)
             sg.delete()
@@ -1895,7 +1909,7 @@ class ClusterValidator(validators.Validator):
                     from_port, to_port,
                     reason="'from_port' must be <= 'to_port'")
             cidr_ip = permission.get('cidr_ip')
-            if not iptools.validate_cidr(cidr_ip):
+            if not iptools.ipv4.validate_cidr(cidr_ip):
                 raise exception.InvalidCIDRSpecified(cidr_ip)
 
     def validate_ebs_settings(self):
